Moved all components that are not generic (which is most of them) to a components folder under the route they're used in.
