+++ src/Routes/AddToCart/Components/AdjustQtyInput/index.tsx
@@ -1,5 +1,5 @@
 import { Container, Stack, IconButton, Typography } from "@mui/material";
-import { AppDispatch } from "../../Redux/store";
+import { AppDispatch } from "../../../../Redux/store";
 import RemoveIcon from "@mui/icons-material/Remove";
 import AddIcon from "@mui/icons-material/Add";
 import React from "react";
