+++ src/Routes/Cart/Components/CartItems/index.tsx
@@ -1,5 +1,5 @@
 import { useSelector } from "react-redux";
-import { RootState } from "../../Redux/store";
+import { RootState } from "../../../../Redux/store";
 import { useNavigate } from "react-router";
 
 //Components
